Logic and SAT Problem
Suppose you have a wedding to plan, and want to arrange the wedding
seating for a
certain number of guests in a hall. The hall has a certain number tables
for seating.
Some pairs of guests are couples or close Friends (F), and want to sit
together at the
same table. Some other pairs of guests are Enemies (E), and must be
separated into
different tables. The rest pairs are Indifferent (I) with each other,
and do not mind sitting
together or not. However, each pair of guests can only have one
relationship, (F), (E) or
(I). You need to find a seating arrangement that satisfies all the
constraints.